+++ src/writers/zarrv2.writer.hh
@@ -25,22 +25,22 @@ struct ZarrV2Writer final : public Writer
   public:
     ZarrV2Writer() = delete;
     ZarrV2Writer(const ImageDims& frame_dims,
-                const ImageDims& tile_dims,
-                uint32_t frames_per_chunk,
-                const std::string& data_root,
-                std::shared_ptr<common::ThreadPool> thread_pool);
+                 const ImageDims& tile_dims,
+                 uint32_t frames_per_chunk,
+                 const std::string& data_root,
+                 std::shared_ptr<common::ThreadPool> thread_pool);
 
     /// Constructor with Blosc compression params
     ZarrV2Writer(const ImageDims& frame_dims,
-                const ImageDims& tile_dims,
-                uint32_t frames_per_chunk,
-                const std::string& data_root,
-                std::shared_ptr<common::ThreadPool> thread_pool,
-                const BloscCompressionParams& compression_params);
+                 const ImageDims& tile_dims,
+                 uint32_t frames_per_chunk,
+                 const std::string& data_root,
+                 std::shared_ptr<common::ThreadPool> thread_pool,
+                 const BloscCompressionParams& compression_params);
     ~ZarrV2Writer() override = default;
 
   private:
-    void flush_() noexcept override;
+    void flush_() override;
 };
 } // namespace acquire::sink::zarr